S3_Facial Landmark Detection_Alignment_Swap/AWS Lamda Files/requirements.txt
requests_toolbelt
opencv-python-headless
dlib==19.21.0
numpy
Pillow==7.2.0